# release-22.1: colexec: memory accounting at row granularity in columnarizer