--- a/Makefile
+++ b/Makefile
@@ -808,7 +808,7 @@ EXECGEN_TARGETS = \
   pkg/sql/colexec/is_null_ops.eg.go \
   pkg/sql/colexec/ordered_synchronizer.eg.go \
   pkg/sql/colexec/quicksort.eg.go \
-  pkg/sql/colexec/rowstovec.eg.go \
+  pkg/sql/colexec/rowtovec.eg.go \
   pkg/sql/colexec/select_in.eg.go \
   pkg/sql/colexec/sort.eg.go \
   pkg/sql/colexec/sorttopk.eg.go \