--- a/pkg/gen/execgen.bzl
+++ b/pkg/gen/execgen.bzl
@@ -78,7 +78,7 @@ EXECGEN_SRCS = [
   "//pkg/sql/colexec:is_null_ops.eg.go",
   "//pkg/sql/colexec:ordered_synchronizer.eg.go",
   "//pkg/sql/colexec:quicksort.eg.go",
-  "//pkg/sql/colexec:rowstovec.eg.go",
+  "//pkg/sql/colexec:rowtovec.eg.go",
   "//pkg/sql/colexec:select_in.eg.go",
   "//pkg/sql/colexec:sort.eg.go",
   "//pkg/sql/colexec:sort_partitioner.eg.go",