--- a/pkg/sql/colexec/columnarizer.go
+++ b/pkg/sql/colexec/columnarizer.go
@@ -19,7 +19,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/colmem"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfra"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
-	"github.com/cockroachdb/cockroach/pkg/sql/rowenc"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/util/buildutil"
@@ -54,16 +53,13 @@ type Columnarizer struct {
 	execinfra.ProcessorBaseNoHelper
 	colexecop.NonExplainable
 
-	mode columnarizerMode
-	// helper is used in the columnarizerBufferingMode mode.
-	helper colmem.AccountingHelper
-	// allocator is used directly only in the columnarizerStreamingMode mode.
-	allocator *colmem.Allocator
-	input     execinfra.RowSource
-	da        tree.DatumAlloc
+	mode   columnarizerMode
+	helper colmem.SetAccountingHelper
+	input  execinfra.RowSource
+	da     tree.DatumAlloc
 
-	buffered        rowenc.EncDatumRows
 	batch           coldata.Batch
+	vecs            coldata.TypedVecs
 	accumulatedMeta []execinfrapb.ProducerMetadata
 	typs            []*types.T
 
@@ -78,6 +74,7 @@ var _ colexecop.VectorizedStatsCollector = &Columnarizer{}
 
 // NewBufferingColumnarizer returns a new Columnarizer that will be buffering up
 // rows before emitting them as output batches.
+// - allocator must a use memory account that is not shared with any other user.
 func NewBufferingColumnarizer(
 	allocator *colmem.Allocator,
 	flowCtx *execinfra.FlowCtx,
@@ -89,6 +86,7 @@ func NewBufferingColumnarizer(
 
 // NewStreamingColumnarizer returns a new Columnarizer that emits every input
 // row as a separate batch.
+// - allocator must a use memory account that is not shared with any other user.
 func NewStreamingColumnarizer(
 	allocator *colmem.Allocator,
 	flowCtx *execinfra.FlowCtx,
@@ -112,12 +110,8 @@ func newColumnarizer(
 		colexecerror.InternalError(errors.AssertionFailedf("unexpected columnarizerMode %d", mode))
 	}
 	c := &Columnarizer{
-		allocator: allocator,
-		input:     input,
-		mode:      mode,
-	}
-	if mode == columnarizerBufferingMode {
-		c.helper.Init(allocator, execinfra.GetWorkMemLimit(flowCtx))
+		input: input,
+		mode:  mode,
 	}
 	c.ProcessorBaseNoHelper.Init(
 		nil, /* self */
@@ -145,6 +139,7 @@ func newColumnarizer(
 			}},
 	)
 	c.typs = c.input.OutputTypes()
+	c.helper.Init(allocator, execinfra.GetWorkMemLimit(flowCtx), c.typs)
 	return c
 }
 
@@ -198,50 +193,22 @@ func (c *Columnarizer) Next() coldata.Batch {
 		return coldata.ZeroBatch
 	}
 	var reallocated bool
-	switch c.mode {
-	case columnarizerBufferingMode:
-		c.batch, reallocated = c.helper.ResetMaybeReallocate(
-			c.typs, c.batch, 0, /* tuplesToBeSet */
-		)
-	case columnarizerStreamingMode:
-		// Note that we're not using ResetMaybeReallocate because we will
-		// always have at most one tuple in the batch.
-		if c.batch == nil {
-			c.batch = c.allocator.NewMemBatchWithFixedCapacity(c.typs, 1 /* minCapacity */)
-			reallocated = true
-		} else {
-			c.allocator.ResetBatch(c.batch)
-		}
+	var tuplesToBeSet int
+	if c.mode == columnarizerStreamingMode {
+		// In the streaming mode, we always emit a batch with at most one tuple,
+		// so we say that there is just one tuple left to be set (even though we
+		// might end up emitting more tuples - it's ok from the point of view of
+		// the helper).
+		tuplesToBeSet = 1
 	}
+	c.batch, reallocated = c.helper.ResetMaybeReallocate(c.typs, c.batch, tuplesToBeSet)
 	if reallocated {
-		oldRows := c.buffered
-		newRows := make(rowenc.EncDatumRows, c.batch.Capacity())
-		copy(newRows, oldRows)
-		if len(oldRows) < len(newRows) {
-			_ = newRows[len(oldRows)]
-			for i := len(oldRows); i < len(newRows); i++ {
-				//gcassert:bce
-				newRows[i] = make(rowenc.EncDatumRow, len(c.typs))
-			}
-		} else if len(newRows) < len(oldRows) {
-			_ = oldRows[len(newRows)]
-			// Lose the reference to the old rows that aren't copied into the
-			// new slice - we need to do this since the capacity of the batch
-			// might have shrunk, and the rows at the end of the slice might
-			// never get overwritten.
-			for i := len(newRows); i < len(oldRows); i++ {
-				//gcassert:bce
-				oldRows[i] = nil
-			}
-		}
-		c.buffered = newRows
+		c.vecs.SetBatch(c.batch)
 	}
-	// Buffer up rows up to the capacity of the batch.
 	nRows := 0
-	for ; nRows < c.batch.Capacity(); nRows++ {
+	for batchDone := false; !batchDone; {
 		row, meta := c.input.Next()
 		if meta != nil {
-			nRows--
 			if meta.Err != nil {
 				// If an error occurs, return it immediately.
 				colexecerror.ExpectedError(meta.Err)
@@ -252,20 +219,11 @@ func (c *Columnarizer) Next() coldata.Batch {
 		if row == nil {
 			break
 		}
-		copy(c.buffered[nRows], row)
+		EncDatumRowToColVecs(row, nRows, c.vecs, c.typs, &c.da)
+		batchDone = c.helper.AccountForSet(nRows)
+		nRows++
 	}
 
-	// Check if we have buffered more rows than the current allocation size
-	// and increase it if so.
-	if nRows > c.da.AllocSize {
-		c.da.AllocSize = nRows
-	}
-
-	// Write each column into the output batch.
-	outputRows := c.buffered[:nRows]
-	for idx, ct := range c.typs {
-		EncDatumRowsToColVec(c.allocator, outputRows, c.batch.ColVec(idx), idx, ct, &c.da)
-	}
 	c.batch.SetLength(nRows)
 	return c.batch
 }
@@ -280,6 +238,9 @@ func (c *Columnarizer) DrainMeta() []execinfrapb.ProducerMetadata {
 	if c.removedFromFlow {
 		return nil
 	}
+	// We no longer need the batch.
+	c.batch = nil
+	c.helper.ReleaseMemory()
 	if c.Ctx == nil {
 		// The columnarizer wasn't initialized, so the wrapped processors might
 		// not have been started leaving them in an unsafe to drain state, so
@@ -303,6 +264,7 @@ func (c *Columnarizer) Close(context.Context) error {
 	if c.removedFromFlow {
 		return nil
 	}
+	c.helper.Release()
 	c.InternalClose()
 	return nil
 }