--- a/pkg/sql/colexec/execgen/cmd/execgen/BUILD.bazel
+++ b/pkg/sql/colexec/execgen/cmd/execgen/BUILD.bazel
@@ -47,7 +47,7 @@ go_library(
         "rank_gen.go",
         "relative_rank_gen.go",
         "row_number_gen.go",
-        "rowstovec_gen.go",
+        "rowtovec_gen.go",
         "select_in_gen.go",
         "selection_ops_gen.go",
         "sort_gen.go",