… pkg/sql/colexec/BUILD.bazel
… pkg/sql/colexec/rowstovec.eg.go
… pkg/sql/colexec/rowstovec_test.go
… pkg/sql/colexec/rowtovec.eg.go
… pkg/sql/colexec/rowtovec_test.go