Backport 2/2 commits from #87242.
Backport 1/1 commits from #87471.

/cc @cockroachdb/release

---

**colexec: remove redundant allocations when converting from rows**

This redundant allocation has been present since the introduction of the
vectorized engine (it was introduced in
37957a9b4e6964940ed2dd4d94b4dbfa671a11aa I think by accident - the
impact of some changes in that commit were probably not understood and
we didn't have the benchmarks to catch the regression).
```
Columnarize-24                442µs ± 1%     157µs ± 0%   -64.49%  (p=0.000 n=10+9)
ColumnarizeMaterialize-24     801µs ± 1%     513µs ± 0%   -35.89%  (p=0.000 n=10+10)

name                       old speed      new speed      delta
Columnarize-24              362MB/s ± 1%  1019MB/s ± 0%  +181.61%  (p=0.000 n=10+9)
ColumnarizeMaterialize-24   200MB/s ± 1%   312MB/s ± 0%   +55.97%  (p=0.000 n=10+10)

name                       old alloc/op   new alloc/op   delta
Columnarize-24                157kB ± 0%       0kB ± 1%   -99.85%  (p=0.000 n=10+9)
ColumnarizeMaterialize-24     356kB ± 0%     200kB ± 0%   -43.94%  (p=0.000 n=10+9)

name                       old allocs/op  new allocs/op  delta
Columnarize-24                19.5k ± 0%      0.0k ± 0%   -99.98%  (p=0.000 n=10+10)
ColumnarizeMaterialize-24     19.5k ± 0%      0.0k ± 0%   -99.77%  (p=0.000 n=10+10)
```
Additionally, the conversion method no longer returns an error that is
encountered during the datum decoding since every caller of the method
is panicking with the error anyway.

Release justification: low-risk improvement.

Release note: None

**colexec: memory accounting at row granularity in columnarizer**

This commit makes it so that we perform the memory accounting at a row
granularity (rather than a batch granularity) in the columnarizer as
well as the values operator. As a result, we will respect memory limits
much better than previously.

The main idea is that previously we were converting from the row-based
format to the columnar one a column at a time, and now we will perform
the conversion one full row at a time. As expected, this leads to
a performance regression, but given the speedup in the previous commit
and the improvement to the stability of Cockroach, it seems worth it.
```
name                       old time/op    new time/op    delta
Columnarize-24                157µs ± 1%     291µs ± 0%  +85.05%  (p=0.000 n=8+10)
ColumnarizeMaterialize-24     514µs ± 0%     648µs ± 0%  +26.09%  (p=0.000 n=9+9)

name                       old speed      new speed      delta
Columnarize-24             1.02GB/s ± 1%  0.55GB/s ± 0%  -45.96%  (p=0.000 n=8+10)
ColumnarizeMaterialize-24   311MB/s ± 0%   247MB/s ± 0%  -20.69%  (p=0.000 n=9+9)

name                       old alloc/op   new alloc/op   delta
Columnarize-24                 229B ± 1%      375B ± 1%  +63.96%  (p=0.000 n=9+10)
ColumnarizeMaterialize-24     200kB ± 0%     200kB ± 0%   +0.05%  (p=0.000 n=10+10)

name                       old allocs/op  new allocs/op  delta
Columnarize-24                 4.00 ± 0%      7.00 ± 0%  +75.00%  (p=0.000 n=10+10)
ColumnarizeMaterialize-24      44.0 ± 0%      47.0 ± 0%   +6.82%  (p=0.000 n=10+10)
```

This commit makes the corresponding change to the method and file names
to indicate that now the conversion is from a single row into all
vectors at once.

Addresses: #76464.

Release justification: bug fix.

Release note: None

----

The performance impact of both commits combined:
```
name                       old time/op    new time/op    delta
Columnarize-24                441µs ± 0%     292µs ± 1%  -33.80%  (p=0.000 n=9+10)
ColumnarizeMaterialize-24     800µs ± 1%     648µs ± 0%  -18.99%  (p=0.000 n=10+10)

name                       old speed      new speed      delta
Columnarize-24              363MB/s ± 0%   548MB/s ± 1%  +51.07%  (p=0.000 n=9+10)
ColumnarizeMaterialize-24   200MB/s ± 1%   247MB/s ± 0%  +23.44%  (p=0.000 n=10+10)

name                       old alloc/op   new alloc/op   delta
Columnarize-24                157kB ± 0%       0kB ± 1%  -99.76%  (p=0.000 n=10+9)
ColumnarizeMaterialize-24     356kB ± 0%     200kB ± 0%  -43.91%  (p=0.000 n=9+9)

name                       old allocs/op  new allocs/op  delta
Columnarize-24                19.5k ± 0%      0.0k ± 0%  -99.96%  (p=0.000 n=10+10)
ColumnarizeMaterialize-24     19.5k ± 0%      0.0k ± 0%  -99.76%  (p=0.000 n=10+9)
```